- updated version example - moved version to lib/version.ts
