--- src/version.ts
@@ -1,8 +1,0 @@
-import { version } from "lib";
-
-if (import.meta.main) {
-  const { micro, major, minor } = version;
-  console.log(`Gtk version: ${major}.${minor}.${micro}`);
-}
-
-export { version };
